% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mehrotra2012.R
\name{CLi_mehrotra}
\alias{CLi_mehrotra}
\title{calculates clearance in L/hr for an individual}
\usage{
CLi_mehrotra(wt, pma, scr)
}
\arguments{
\item{wt}{weight in kg}

\item{pma}{postmentstral age, weeks}

\item{scr}{serum creatinine mg/dl}
}
\description{
calculates clearance in L/hr for an individual
}
\details{
Mehrotra, N., Tang, L., Phelps, S. J. & Meibohm, B. Evaluation of vancomycin
dosing regimens in preterm and term neonates using Monte Carlo simulations.
Pharmacotherapy 32, 408–419 (2012)
}
