% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockman2015.R
\name{CLi_stockmann2015}
\alias{CLi_stockmann2015}
\title{calculates clearance in L/hr for an individual}
\usage{
CLi_stockmann2015(wt, pma, scr)
}
\arguments{
\item{wt}{weight in kg}

\item{pma}{postmentstral age, weeks}

\item{scr}{serum creatinine mg/dl}
}
\description{
calculates clearance in L/hr for an individual
}
\details{
Stockmann, C. et al. Predictive Performance of a Vancomycin
Population Pharmacokinetic Model in Neonates.
Infect Dis Ther 4, 187–198 (2015).
}

