% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2bin_pna_scr_wt.R
\name{dosing_2bin_pna_scr_wt}
\alias{dosing_2bin_pna_scr_wt}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_2bin_pna_scr_wt(pna, scr, wt)
}
\arguments{
\item{pna}{post natal age (days)}

\item{scr}{serum creatinine umol/L}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
as used in publications Anderson et al, 2007; Allegaert et al., 2007;
kimura et al., 2004 -
PNA 1-7 days, serum creatinine > 71 umol/L: 15 mg/kg every 12 hrs;
PNA > 7 days: 15 mg/kg every 8 hrs
}

