% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4bin_pma_wt.R
\name{dosing_4bin_pma_wt}
\alias{dosing_4bin_pma_wt}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_4bin_pma_wt(pma, wt)
}
\arguments{
\item{pma}{post menstrual age (weeks)}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
DR2
}
