% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4bin_pna_wt.R
\name{dosing_4bin_pna_wt}
\alias{dosing_4bin_pna_wt}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_4bin_pna_wt(pna, wt)
}
\arguments{
\item{pna}{post natal age (days)}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
as used in publication Madigan, Theresa et al. "Optimization of Vancomycin 
Dosing in Very Low-Birth-Weight Preterm Neonates." 
American journal of perinatology 32.1 (2015): 83–86. PMC. Web. 27 Mar. 2016.
}
