% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5bin_scr_wt.R
\name{dosing_5bin_scr_wt}
\alias{dosing_5bin_scr_wt}
\title{dosing nomogram based on serum creatinine}
\usage{
dosing_5bin_scr_wt(scr, wt)
}
\arguments{
\item{scr}{serum creatinine (mg/dL)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on serum creatinine
}
\details{
From the Red Book 2012
\tabular{lll}{
serum creatinine \tab dose \tab interdose interval
mg/dL   \tab mg/kg \tab hr\cr
< 0.7   \tab 15    \tab 12\cr
0.7-0.9 \tab 20    \tab 24\cr
1-1.2   \tab 15    \tab 24\cr
1.3-1.6 \tab 10    \tab 24\cr
> 1.6   \tab 15    \tab 48
}
}

