% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6bin_pma_pna_wt.R
\name{dosing_6bin_pma_pna}
\alias{dosing_6bin_pma_pna}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_6bin_pma_pna(pma, pna, wt)
}
\arguments{
\item{pma}{post menstrual age (weeks)}

\item{pna}{post natal age (days)}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
DR3
}
