% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6bin_pna_wt.R
\name{dosing_6bin_pna_wt}
\alias{dosing_6bin_pna_wt}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_6bin_pna_wt(pna, wt)
}
\arguments{
\item{pna}{post natal age (days)}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
as used in publication Dersch-Mills,et al.,
 Assessment of initial vancomycin dosing in neonates. 
 Paediatr Child Health 19, e30–e34 (2014).

recommended algorithm from Nelson's textbook of Pediatrics 2007
}

