% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7bin_pma_pna_wt.R
\name{dosing_7bin_pma_pna_wt}
\alias{dosing_7bin_pma_pna_wt}
\title{dosing nomogram based on post natal age and serum creatinine}
\usage{
dosing_7bin_pma_pna_wt(pma, pna, wt)
}
\arguments{
\item{pma}{post menstrual age (weeks)}

\item{pna}{post natal age (days)}

\item{wt}{weight (kg)}
}
\value{
dataframe with DOSE and II columns
}
\description{
dosing nomogram based on post natal age and serum creatinine
}
\details{
Rajon et al 2017
}
