% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mehrotra2012.R
\name{Vi_mehrotra}
\alias{Vi_mehrotra}
\title{calculates Vi L for an individual}
\usage{
Vi_mehrotra(wt)
}
\arguments{
\item{wt}{weight in kg}
}
\description{
calculates Vi L for an individual
}
\details{
Mehrotra, N., Tang, L., Phelps, S. J. & Meibohm, B. Evaluation of vancomycin
dosing regimens in preterm and term neonates using Monte Carlo simulations.
Pharmacotherapy 32, 408–419 (2012)
}
