% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockman2015.R
\name{Vi_stockmann2015}
\alias{Vi_stockmann2015}
\title{calculates Vi L for an individual}
\usage{
Vi_stockmann2015(wt)
}
\arguments{
\item{wt}{weight in kg}
}
\description{
calculates Vi L for an individual
}
\details{
Stockmann, C. et al. Predictive Performance of a Vancomycin 
Population Pharmacokinetic Model in Neonates. 
Infect Dis Ther 4, 187–198 (2015).
}

